% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apricitas.R
\name{theme_apricitas}
\alias{theme_apricitas}
\title{Define the theme of the Apricitas blog}
\usage{
theme_apricitas()
}
\description{
Define the theme of the Apricitas blog
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_apricitas()
}
